requirements.txt
-r flomo/requirements.txt

click
imapclient
python-dotenv
PyYAML
